+++ buildSrc/src/main/kotlin/ktx/Versions.kt
@@ -1,12 +1,12 @@
 package ktx
 
-const val gdxVersion = "1.12.0"
-const val kotlinCoroutinesVersion = "1.7.2"
+const val gdxVersion = "1.12.1"
+const val kotlinCoroutinesVersion = "1.7.3"
 
 const val artemisOdbVersion = "2.3.0"
 const val ashleyVersion = "1.7.4"
 const val gdxAiVersion = "1.8.2"
-const val visUiVersion = "1.5.2"
+const val visUiVersion = "1.5.3"
 
 const val spekVersion = "1.1.5"
 const val kotlinTestVersion = "2.0.7"